Use reflection extensions from RapidCore
Even though the InvokeMethodRecursively (and friends) method does not support taking in a null parameter (for the moment), this still makes the codebase more robust. The regular built-in methods for getting methods, do not search through the inheritance tree, but the new extensions do, which makes our code more resilient to implementations moving around.