Add support for indexes on nested types
As an entity might not have any indexes directly, but it might have indexes on nested types, I needed a way to to tell the system to include a type in the index scanning. I chose to do this via the new [Entity] attribute. Before we simply included types that had at least one [Index].