Added abstractions for acquiring distributed application locks
This commit adds pure interfaces for this, as we have no real distributed
app lock to implement here, as it would pull in too many external libraries.

One acquires the lock by creating a new concrete instance of an `IDistributedAppLocker`
which upon acquisition of a lock returns a `IDistributedAppLock` which must be disposed
in order to release the lock handle.

The idea for providers of a Distributed App Lock is that they _must_ guarantee to lock
across all IIS AppPools or dotnet processes running, which are connected to the same
locking coordinator and trying to acquire a lock of the same name. Lock coordinators could
be SQL server, Redis or Apache ZooKeeper.