Return extenable model instead of bool from On* methods
Instead of returning a boolean with a meaning that is not obvious from
the call tree or the name of things, we now return
IInstanceListenerOnFieldOrPropResult. Currently it only has a property
for the boolean value, but the idea is that people will be able to
add their own things to the output.