Remove SSH.NET and the IFileSystemProvider
This _should_ have been in its own package (as it introduced an external
dependency). It has now caused problems for someone down the line due to
the dependencies of SSH.NET itself. We are not aware of anyone using
this feature, so we are choosing the simple option and removing it
completely.